Abort on panics in futures
Without an explicit handler, it's not clear what to do if there's a
Rust panic in the fulfillment of a JavaScript future. Older versions
of Node silently drop those failures, so to make them "loud" again the
code here will abort on an unhandled unwinding panic.

libsignal-node will use its own panic handler to avoid crashing the
process. This is just to keep issues from going unnoticed.