Panic on unhandled JavaScript exceptions in async callbacks
Similar to the previous commit, we don't want to allow JavaScript
exceptions to escape into the microtask queue, because they'll be
silently dropped. Convert them to panics so that the logic from the
previous commit can catch them.

Also, add a utility to make sure we print these panics correctly.